auth password hash jwt tokens

+++ src/services/user.service.ts
@@ -1,4 +1,10 @@
-import { IUser, UserDTO } from "../interfaces/user.interface";
+import { StatusCodes } from "../enums/status-codes.enum";
+import { ApiError } from "../errors/api.error";
+import {
+    IUser,
+    UserCreateDTO,
+    UserUpdateDTO,
+} from "../interfaces/user.interface";
 import { userRepository } from "../repositories/user.repository";
 
 class UserService {
@@ -6,20 +12,52 @@ class UserService {
         return userRepository.getAllUsers();
     }
 
-    public createUser(user: UserDTO): Promise<IUser> {
+    public createUser(user: UserCreateDTO): Promise<IUser> {
         return userRepository.createUser(user);
     }
 
-    public getUserById(userId: string): Promise<IUser | null> {
-        return userRepository.getUserById(userId);
+    public async getUserById(userId: string): Promise<IUser> {
+        const user = await userRepository.getUserById(userId);
+
+        if (!user) {
+            throw new ApiError("User not found", StatusCodes.NOT_FOUND);
+        }
+
+        return user;
     }
 
-    public updateUserById(userId: string, data: IUser): Promise<IUser | null> {
-        return userRepository.updateUserById(userId, data);
+    public async updateUserById(
+        userId: string,
+        user: UserUpdateDTO,
+    ): Promise<IUser> {
+        const data = await userRepository.updateUserById(userId, user);
+
+        if (!data) {
+            throw new ApiError("User not found", StatusCodes.NOT_FOUND);
+        }
+
+        return await userRepository.updateUserById(userId, user);
     }
 
-    public deleteUserById(userId: string): Promise<IUser | null> {
-        return userRepository.deleteUserById(userId);
+    public async deleteUserById(userId: string): Promise<void> {
+        const data = await userRepository.deleteUserById(userId);
+
+        if (!data) {
+            throw new ApiError("User not found", StatusCodes.NOT_FOUND);
+        }
+
+        await userRepository.deleteUserById(userId);
+    }
+
+    public async isEmailUnique(email: string): Promise<void> {
+        const user = await userRepository.getUserByEmail(email);
+
+        if (user) {
+            throw new ApiError(
+                "User is already exists",
+                StatusCodes.BAD_REQUEST,
+            );
+        }
     }
 }
